✨[Feat] 대기 승인 뷰에 로그아웃 및 회원 탈퇴 기능 추가
- FailedVerificationUMCViewModel 추출 (View → ViewModel 로직 분리)
- 로그아웃 기능 구현 (presentLogoutPrompt + logout 메서드)
- 기존 계정 삭제 기능을 ViewModel로 이관
- 하단 툴바 "계정" 버튼을 Menu로 변경 (로그아웃/회원 탈퇴 선택)
- ToolBarCollection.FailedVerificationBottomToolbar에 onLogout, isLoggingOut 추가
- 테스트 파일 작성자 정보 수정
