+++ AppProduct/AppProductTests/AuthTest/SignUpViewModelTests.swift
@@ -2,7 +2,7 @@
 //  SignUpViewModelTests.swift
 //  AppProductTests
 //
-//  Created by Codex on 3/10/26.
+//  Created by euijjang97 on 3/10/26.
 //
 
 @testable import AppProduct
