+++ AppProduct/AppProductTests/HomeTest/SearchChallengerViewModelTests.swift
@@ -2,7 +2,7 @@
 //  SearchChallengerViewModelTests.swift
 //  AppProductTests
 //
-//  Created by Codex on 3/10/26.
+//  Created by euijjang97 on 3/10/26.
 //
 
 @testable import AppProduct
